Starting experimental work on 4.0.0
This commit represents starting work on Reactor Core generation 4.

The `gen4` branch will contain commits leading up to 4.0.0 GA release,
at most. We expect `main` to at some point be fast-forwarded to `gen4`
at which time the effort will continue on main, and latest 3.x.y
development will have its own separate branch.

In the meantime, we expect regular rebasing of this branch on top of
`main`, to prepare and allow for that later fast-forward.